scripts/requirements.txt
SQLAlchemy==2.0.22
psycopg2-binary==2.9.3
confluent-kafka==2.3.0